requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
requests
beautifulsoup4
pandas
matplotlib
scikit-learn
lxml